backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
celery>=5.3.0
redis>=5.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
asgiref>=3.7.0
openpyxl>=3.1.0
reportlab>=4.0.0
